Refactor ThomasModelService
