+++ src/main/java/fiuba/tpp/reactorapp/service/MathService.java
@@ -4,8 +4,6 @@ import fiuba.tpp.reactorapp.entities.Process;
 import fiuba.tpp.reactorapp.model.exception.NotEnoughObservationsException;
 import fiuba.tpp.reactorapp.model.math.Observation;
 import fiuba.tpp.reactorapp.model.math.RegressionResult;
-import fiuba.tpp.reactorapp.model.request.ThomasRequest;
-import fiuba.tpp.reactorapp.model.response.ThomasResponse;
 import org.apache.commons.math3.stat.regression.SimpleRegression;
 import org.apache.commons.math3.util.Precision;
 import org.springframework.stereotype.Service;
@@ -57,28 +55,7 @@ public class MathService {
         return Math.log(value);
     }
 
-    public ThomasResponse calculateThomas(RegressionResult regression, ThomasRequest request){
-        double kth = round(thomasConstant(regression,request));
-        double qo = round(thomasQo(regression,request,kth));
-
-        return new ThomasResponse(kth,qo);
-    }
-
-    //-Kth * Co / F = slope
-    // Entonces
-    // Kth =- slope * F /Co
-    private double thomasConstant(RegressionResult regression, ThomasRequest request){
-        return -(regression.getSlope() * request.getCaudalVolumetrico()) / request.getConcentracionInicial();
-    }
-
-    //Una vez que tenemos Kth
-    //Kth * Qo * W /F = intercept
-    // Qo = intercept * F / W * Kth
-    private double thomasQo(RegressionResult regression, ThomasRequest request, double thomasConstant){
-        return (regression.getIntercept() * request.getCaudalVolumetrico()) / (request.getSorbenteReactor() * thomasConstant);
-    }
-
-    private double round(double value){
+    public double round(double value){
         return Precision.round(value,NUMBER_OF_DECIMALS);
     }
 }
